% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcCorr}
\alias{calcCorr}
\title{Calculate pairwise correlations and print to a file.}
\usage{
calcCorr(fls, MaxNameLen, recSize, P, n, zeroSD = 0.001)
}
\arguments{
\item{fls}{A character vector which includes the names of the input and output files.}

\item{MaxNameLen}{The length of the first column in the input data file (predictor names).}

\item{recSize}{The total length of input rows (fixed-width format).}

\item{P}{The number of predictors.}

\item{n}{The sample size.}

\item{zeroSD}{If the standard deviation of a column is below this threshold, this column will be considered as uncorrelated with all other columns (Default=1e-3).}
}
\description{
Calculate pairwise correlations and print to a file.
}
