% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMix.R
\name{clusteringCoef}
\alias{clusteringCoef}
\title{Calculate the clustering coefficient of each node.}
\usage{
clusteringCoef(A)
}
\arguments{
\item{A}{an adjacency Matrix (0/1).}
}
\value{
A vector with the clustering coefficient of each node.
}
\description{
Calculate the clustering coefficient of each node.
}
\examples{
\donttest{
   data(SIM)
   res <- betaMix(SIM, delta = 1e-6,ppr = 0.01,subsamplesize = 30000, ind=TRUE)
   adjMat <- getAdjMat(res)
   clusteringCoef(adjMat)
}

}
