% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMix.R
\docType{data}
\name{DrySeeds}
\alias{DrySeeds}
\title{Metabolite Expression data for the the dry seed group.}
\format{
A matrix with 50 rows and 68 columns
}
\description{
DrySeeds is a matrix with normalized metabolite data containing with 68 named metabolites, and 50 samples.
}
\references{
\url{https://www.metabolomicsworkbench.org/data/DRCCMetadata.php?Mode=Study&StudyID=ST000561&StudyType=MS&ResultType=1}
}
\keyword{datasets}
