% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMix.R
\name{plotBitmapCC}
\alias{plotBitmapCC}
\title{Edge-indicator bitmap plot.}
\usage{
plotBitmapCC(
  AdjMat,
  clusterInfo = NULL,
  orderByCluster = FALSE,
  showMinDegree = 0
)
}
\arguments{
\item{AdjMat}{An adjacency Matrix (0/1).}

\item{clusterInfo}{obtained from graphComponents. If not provided by the user, it will be computed on the fly.}

\item{orderByCluster}{If \code{FALSE}, show the bitmap in the original node
order.  If \code{TRUE}, show nodes by clusters, sorted by distance from
the cluster center.  Note: if \code{clusterInfo} is provided this
argument is ignored and cluster ordering is always applied.}

\item{showMinDegree}{Non-negative integer indicating the minimum degree of nodes that should be displayed. Default=0 (all nodes).}
}
\description{
Plot a bitmap in which a black dot corresponds to a pair of highly correlated genes (an edge in the graph).
The default is to show the nodes according to their order in the input.
By setting orderByCluster=TRUE, it is possible to cluster nodes and show
them in increasing degree order (from left to right).
}
\examples{
\dontrun{
data(SIM, package = "betaMix")
res <- betaMix(betaMix::SIM, maxalpha = 1e-6, ppr = 0.01, subsamplesize = 30000, ind = TRUE)
adjMat <- getAdjMat(res)
SimComp <- graphComponents(adjMat)
plotBitmapCC(adjMat)
plotBitmapCC(adjMat, SimComp, orderByCluster = TRUE)
}
}
\seealso{
\code{\link{betaMix-package}} for a package overview;
  \code{help(package = "betaMix")} to browse the full function and dataset index.
}
