% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMix.R
\docType{data}
\name{SIM}
\alias{SIM}
\title{Simulated gene Expression data using the huge package}
\format{
A matrix with 200 rows (samples) and 1000 columns (variables/nodes), representing 50 hubs
}
\usage{
data(SIM,package = "betaMix")
}
\description{
SIM is a simulated dataset with a hub structure, consisting of 1000 nodes (variables),
200 samples, and 50 hubs.
}
\keyword{datasets}
