% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMix.R
\name{sphericalCaps}
\alias{sphericalCaps}
\title{Find hub-like subgraphs (spherical caps) in the network.}
\usage{
sphericalCaps(A)
}
\arguments{
\item{A}{An adjacency Matrix(0/1), zeros on the main diagonal.}
}
\value{
A data frame with the following columns:
\describe{
\item{node}{Node number (row/column in A.)}
\item{capNum}{The cap number.}
\item{isCtr}{1 if the node is the center of the cap, 0 otherwise.}
\item{deg}{Node degree.}
\item{cc}{Clustering coefficient.}
}
}
\description{
Identifies hub structures in the network graph by greedily selecting
high-degree nodes as cap centers.  Each center and all its direct neighbors
form a "cap".  Centers are chosen in decreasing degree order and each node
can serve as a center at most once; however, non-center nodes may appear in
multiple caps if they are neighbors of several centers.  Conceptually, each
hub corresponds to a spherical cap in the high-dimensional space underlying
the beta-mixture model.
}
\examples{
\dontrun{
data(SIM, package = "betaMix")
res <- betaMix(betaMix::SIM, maxalpha = 1e-6, ppr = 0.01, subsamplesize = 30000, ind = TRUE)
adjMat <- getAdjMat(res)
caps <- sphericalCaps(adjMat)
head(caps)
am <- getAdjMat(res, signed = TRUE, nodes = caps$node[which(caps$capNum == 1)])
plotCluster(am, 1, edgecols = c("blue", "red"), labels = TRUE)
}
}
\seealso{
\code{\link{betaMix-package}} for a package overview;
  \code{help(package = "betaMix")} to browse the full function and dataset index.
}
