% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMix.R
\name{getAdjMat}
\alias{getAdjMat}
\title{Get the adjacency matrix if the pairwise correlation data is stored in a SQL database.}
\usage{
getAdjMat(res, dbname = NULL, ppthr = NULL, signed = FALSE, nodes = NULL)
}
\arguments{
\item{res}{The returned value from the betaMix function if the correlation matrix is strored in memory.}

\item{dbname}{The sqlite database with the pairwise correlations, if the data is not stored in memory (for very large P). Default=NULL.}

\item{ppthr}{A threshold to select edges. Default=NULL, in which case the returned value from betaMix() is used (available in res).}

\item{signed}{If TRUE, the returned matrix will contain -1 for negatively correlated pairs. Otherwise, all correlated pairs will have 1 in the returned matrix (Default).}

\item{nodes}{An optional parameter which allows to create a sparse adjacency matrix containing only the neighbors of the selected nodes (default=NULL).}
}
\description{
SQL storage should be used if P is large and the calculating and storing the correlation matrix will require too much memory.
}
\examples{
\donttest{
   res <- betaMix(SIM, delta = 1e-5,ppr = 0.01,subsamplesize = 30000, ind=TRUE)
   adjMat <- getAdjMat(res)
   image(adjMat[1:80,1:80])
   # See online documentation for usage when SQLite is used.
}
}
