% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMix.R
\name{plotDegCC}
\alias{plotDegCC}
\title{Plot the degree of nodes versus the degree times the clustering coefficient.}
\usage{
plotDegCC(betamixobj, clusterInfo = NULL, highlightNodes = NULL)
}
\arguments{
\item{betamixobj}{The object (list) returned by betaMix}

\item{clusterInfo}{obtained from graphComponents. If not provided by the user, it will be computed on the fly.}

\item{highlightNodes}{A vector of node-numbers which will be shown in red. Default is NULL.}
}
\description{
The x-axis represents the number of neighbors of each node, and the y-axis represents the proportion of neighbors which are connected to each other.
}
\examples{
\dontrun{
   data(SIM,package = "betaMix")
   res <- betaMix(betaMix::SIM, maxalpha = 1e-6,ppr = 0.01,subsamplesize = 30000, ind=TRUE)
   adjMat <- getAdjMat(res)
   SimComp <- graphComponents(adjMat)
   plotDegCC(res,SimComp)
}
}
