% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMix.R
\name{plotFittedBetaMix}
\alias{plotFittedBetaMix}
\title{Plot the histogram of the z_j and the fitted mixture distribution.}
\usage{
plotFittedBetaMix(betaMixObj, yLim = 5)
}
\arguments{
\item{betaMixObj}{An object returned from betaMix()}

\item{yLim}{The maximum value on the y-axis (default=5)}
}
\description{
Plot the histogram of the z_j and the fitted mixture distribution.
}
\examples{
\donttest{
   data(SIM,package = "betaMix")
   res <- betaMix(betaMix::SIM, delta = 1e-5,ppr = 0.01,subsamplesize = 30000, ind=TRUE)
   plotFittedBetaMix(res)
}
}
