% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMix.R
\name{shortSummary}
\alias{shortSummary}
\title{Print a short summary of the fitted mixture model.}
\usage{
shortSummary(betamixobj)
}
\arguments{
\item{betamixobj}{The object (list) returned from the betaMix function.}
}
\description{
Show the number of nodes, the number of possible and detected edges, the estimated proportion of uncorrelated pairs.
}
\examples{
\donttest{
   data(SIM)
   res <- betaMix(SIM, delta = 1e-6,ppr = 0.01,subsamplesize = 30000, ind=TRUE)
   shortSummary(res)
}
}
